Use LZ4 block compression for compatibility with parquet-cpp (#314)
This removes LZ4 frame compression in favour of LZ4 block
compression to ensure the LZ4 compression used is compatible
with the paqrquet-cpp implementation.

Because the lz4.block decompressor needs to know the size
of the uncompressed data, this change also requires:

    - propagation of page_header.uncompressed_page_size
      through to decompress_data
    - the decompression functions for all compressors
      to handle the uncompressed_size argument

Where a decompressor doesn't need to know the uncompressed
data size, we simply ignore it.