Persist and read column index pandas metadata
This is to address issue #409, in which column index information (e.g.,
name of the index) had not previously been persisted in the pandas
metadata dictionary.  This commit borrows code from pyarrow to both
write this metadata (the 'column_indexes' field) and to use it in
loading.  Note that this commit does not fix the issue also mentioned
in #409 still breaks upon trying to write a dataframe with a multi-level
column index.