enable multi-level column index write/read
This enables dataframes with multi-level column indexes
to be both written and read.  It also gets rid of a test
that I'm not sure is relevant anymore---that is, the test was
asking for an exception to be raised, but the code now lets it run
(maybe this is a problem?).  The test in question is test_bad_coltype
in test_output.py.